🚀 feat(github): add GitHub Actions workflow for building and deploying the website 📄 feat(articles): create initial articles for the website
A new GitHub Actions workflow is added to automate the build and deployment of the website upon pushes and pull requests to the main branch. This enhances the CI/CD process, ensuring that the latest changes are automatically built and deployed. Additionally, initial articles are created, including a home page and a page on neural networks, to provide content for the website and engage users.

📄 feat(oldindex.md): add new article "The Monospace Web" exploring minimalist design with monospace fonts
🖼️ feat(castle.jpg): add image of a room in an old French castle to enhance the article

The new article discusses the aesthetic and technical aspects of using monospace fonts on the web. It includes various examples of HTML elements styled in a monospace grid layout, providing a creative exploration of design. The accompanying image adds visual interest and context to the content, enriching the reader's experience.